updated project and worked on the search system for the home page

+++ lib/src/repo/models/rating/ratings.dart
@@ -9,7 +9,7 @@ import '../../utils/helpers.dart';
 
 class Ratings {
   final String? id;
-  final int? ratingValue;
+  final double? ratingValue;
   final String? comment;
   final DateTime? created;
   final User client;
